projects/wmflex/requirements.txt
pillow
psutil